requirements.txt
torch>=2.6.0
torchvision>=0.21.0
transformers
accelerate
diffusers
ftfy
tqdm
onnx
optuna
datasets>=3.2.0
pycocotools~=2.0.8
numpy
pandas
scipy
scikit-learn
requests>=2.32.3
pillow
setuptools
nltk>=3.9.1
gdown

# For the code evaluator
flake8~=7.1.1
flake8-json~=24.4.0
radon~=6.0.1
pylint~=3.3.3
git+https://github.com/salaniz/pycocoevalcap
